refactor: model -> entity 패키지명 수정
